backend/todos/views.py
from django.shortcuts import render
from rest_framework import generics
from .serializers import TodoSerializer
from .models import Todo

# function based imports
from rest_framework.response import Response
from rest_framework.decorators import api_view
from rest_framework import status

class ListTodo(generics.ListAPIView):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer

class DetailTodo(generics.RetrieveAPIView):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer    

@api_view(['GET'])
def list_view(request):
    todos = Todo.objects.all()
    serializer = TodoSerializer(todos, many=True)
    return Response(serializer.data)

@api_view()
def detail_view(request, pk):
    try:
        todo = Todo.objects.get(pk=pk)
    except Todo.DoesNotExist:
        return Response(status=status.HTTP_404_NOT_FOUND)
    serializer = TodoSerializer(todo)
    return Response(serializer.data)
