backend/requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
pytz==2020.1
sqlparse==0.3.1